pyproject.toml
[tool.poetry]
name = "telegram-bot-demo"
version = "0.1.0"
description = ""
authors = ["dmitriy.balashov <reaver_std@mail.ru>"]
readme = "README.md"
packages = [{include = "telegram_bot_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.12.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
